requirements.txt
beautifulsoup4
Flask
line-bot-sdk
python-firebase
config
imgurpython
SpeechRecognition
pydub
gunicorn==19.7.1
selenium==3.8.0
pyquery
pytesseract
pillow




